viewer/backend/requirements.txt
Django>=5.0
djangorestframework>=3.15.1
django-cors-headers>=4.3.1
django-filter>=24.2
djangorestframework-simplejwt>=5.3.1
psycopg2-binary>=2.9.9
celery>=5.4.0
redis>=5.0.4
python-dotenv>=1.0.1
pillow>=10.3.0
django-cleanup>=8.1.1
django-extensions>=3.2.3
requests>=2.31.0
